2021.1.25 Finish modify project.

+++ app/src/main/java/com/nano/activity/devicechoose/DeviceChooseFragment.java
@@ -82,10 +82,10 @@ public class DeviceChooseFragment  extends Fragment {
         // 说明找到了
         if (historyIndex != PersistUtil.DEFAULT_INTEGER_VALUE) {
             spServiceLife.setSelectedIndex(historyIndex);
-            device.setServiceLife(historyIndex);
+            device.setServiceLife((double)historyIndex);
         } else {
             spServiceLife.setSelectedIndex(0);
-            device.setServiceLife(0);
+            device.setServiceLife(0D);
         }
 
         etSerialNumber.addTextChangedListener(new TextWatcher() {
@@ -106,7 +106,7 @@ public class DeviceChooseFragment  extends Fragment {
             @Override
             public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
                 int serviceLife = spServiceLife.getSelectedIndex();
-                device.setServiceLife(serviceLife);
+                device.setServiceLife((double)serviceLife);
                 PersistUtil.putIntegerValue("DeviceServiceLifeIndex" + device.getDeviceCode(), serviceLife);
             }
             @Override
